+++ src/main/java/com/pbs/ams/web/mappers/AmsTradeInstructMapper.java
@@ -3,13 +3,13 @@ package com.pbs.ams.web.mappers;
 import com.pbs.ams.web.model.AmsTradeInstruct;
 
 public interface AmsTradeInstructMapper {
-    int deleteByPrimaryKey(Integer instructId);
+    int deleteByPrimaryKey(Long instructId);
 
     int insert(AmsTradeInstruct record);
 
     int insertSelective(AmsTradeInstruct record);
 
-    AmsTradeInstruct selectByPrimaryKey(Integer instructId);
+    AmsTradeInstruct selectByPrimaryKey(Long instructId);
 
     int updateByPrimaryKeySelective(AmsTradeInstruct record);
